requirements.txt
certifi==0.0.8
chardet==1.0.1
distribute==0.6.24
oauthlib==0.1.3
pyasn1==0.1.3
requests==0.13.0
rsa==3.0.1
pyyaml==3.10
